requirements.txt
Flask>=2.0
pillow>=9.0
pytesseract>=0.3
Werkzeug>=2.0
opencv-python-headless>=4.5
